requirements.txt
asgiref==3.5.2
black==22.10.0
click==8.1.3
Django==4.1.3
djangorestframework==3.14.0
mypy-extensions==0.4.3
pathspec==0.10.2
platformdirs==2.5.4
pytz==2022.6
sqlparse==0.4.3
tomli==2.0.1
typing-extensions==4.4.0
psycopg2-binary